% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scrapRSS_func.R
\name{scrap_RSS}
\alias{scrap_RSS}
\title{Scrapea el RSS de un periódico y devuelve un json con las noticas del día actual.}
\usage{
scrap_RSS(url)
}
\arguments{
\item{url}{}
}
\value{
json
}
\description{
Recibe la URL RSS de un periódico, lo scrapea, devuelve un json con las noticas del día actual y lo envía a la plataforma Smart City.
}
\examples{
 scrap_RSS('https://www.deia.eus/rss/economia.xml')

}
